refactor: create a /trasnforms folder
#58
