+++ lib/eleventy.config.js
@@ -12,7 +12,9 @@ module.exports = function (eleventyConfig) {
   eleventyConfig.addFilter('hostname', require('./filters/hostname.js'));
   eleventyConfig.addFilter('strip_mentions', require('./filters/strip-mentions.js'));
   eleventyConfig.addFilter('webmentions_for_url', require('./filters/webmentions-for-url.js'));
-  eleventyConfig.addTransform("htmlmin", require('./filters/htmlmin.js'));
+
+  // Transforms
+  eleventyConfig.addTransform("htmlmin", require('./transforms/htmlmin.js'));
 
   // Plugins
   eleventyConfig.addPlugin(require('eleventy-plugin-vimeo-embed'), {
